Add eslint junit report to CircleCI build.
CircleCI supports reading `junit`-formatted reports from tests (or in
this case eslint) and generates nicely formatted error messages rather
than having the cli output be the only indication of what failed.

Because eslint will only output the one format (it won't generate human
readable output when formatting for junit) I added an `on_fail` step
that will generate some human-readable output. That may not be
necessary, since the CircleCI test result output might be the best way
of viewing the information, but it doesn't hurt to have a backup.