Add missing authentication step to publish package job.
Before `npm publish` can be run, we need to be authenticated. The
easiest way to configure CircleCI to be authenticated with npm is by
using an environment variable passed in through the organization
context.  The only thing we need is a valid authentication token which
can be placed in an `.npmrc` file to make it easy to publish.